Include SystemInfoService extension to its own class
