+++ AEPCampaignClassic/Sources/Extensions/SystemInfoService+CampaignClassic.swift
@@ -1,0 +1,28 @@
+/*
+ Copyright 2022 Adobe. All rights reserved.
+ This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ you may not use this file except in compliance with the License. You may obtain a copy
+ of the License at http://www.apache.org/licenses/LICENSE-2.0
+
+ Unless required by applicable law or agreed to in writing, software distributed under
+ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ OF ANY KIND, either express or implied. See the License for the specific language
+ governing permissions and limitations under the License.
+ */
+
+import Foundation
+
+extension SystemInfoService {
+    /// Gets the formatted locale
+    /// - Return: `String` formatted locale
+    func getFormattedLocale() -> String {
+        let locale = getActiveLocaleName()
+        return locale.replacingOccurrences(of: "_", with: "-")
+    }
+
+    /// Gets the formatted operating system containing OSName and OSVersion
+    /// - Return: `String` formatted operating system
+    func getFormattedOperatingSystem() -> String {
+        return "\(getOperatingSystemName()) \(getOperatingSystemVersion())"
+    }
+}
